Make the ITMS Project runnable

+++ db/migrate/20250120073110_drop_user_notifications.rb
@@ -1,5 +1,10 @@
 class DropUserNotifications < ActiveRecord::Migration[7.1]
   def change
-    drop_table :user_notifications
+    # Check if the table exists before trying to drop it
+    if table_exists?(:user_notifications)
+      drop_table :user_notifications
+    else
+      puts "Table 'user_notifications' does not exist, skipping drop."
+    end
   end
 end
